Update pl46  .py

+++ pl46  .py
@@ -3,7 +3,7 @@ A=int(input())
 b=math.radians(A)
 c=math.sin(b)
 if(b<1):
-    print(int(round(c,1)))
+    print(round(c,1))
 elif(b>1):
     print(round(c))
